♻️ Move _flatten_hierarchy_dict to hierarchy.Hierarchy
